fix(xrpl): missing renamings for request package
