+++ xrpl/model/requests/account/account_currencies_response.go
@@ -1,7 +1,7 @@
 package account
 
 import (
-	"github.com/Peersyst/xrpl-go/xrpl/model/client/common"
+	"github.com/Peersyst/xrpl-go/xrpl/model/requests/common"
 )
 
 type AccountCurrenciesResponse struct {
