+++ xrpl/model/requests/admin/data/can_delete_request.go
@@ -3,7 +3,7 @@ package data
 import (
 	"encoding/json"
 
-	"github.com/Peersyst/xrpl-go/xrpl/model/client/common"
+	"github.com/Peersyst/xrpl-go/xrpl/model/requests/common"
 )
 
 type CanDeleteRequest struct {
