+++ xrpl/model/requests/admin/data/download_shard_request.go
@@ -1,6 +1,6 @@
 package data
 
-import "github.com/Peersyst/xrpl-go/xrpl/model/client/common"
+import "github.com/Peersyst/xrpl-go/xrpl/model/requests/common"
 
 type DownloadShardRequest struct {
 	Shards []ShardDescriptor `json:"shards"`
