+++ xrpl/model/requests/admin/data/ledger_request_request.go
@@ -1,6 +1,6 @@
 package data
 
-import "github.com/Peersyst/xrpl-go/xrpl/model/client/common"
+import "github.com/Peersyst/xrpl-go/xrpl/model/requests/common"
 
 type LedgerRequestRequest struct {
 	LedgerIndex common.LedgerIndex `json:"ledger_index,omitempty"`
