+++ xrpl/model/requests/admin/server/ledger_accept_response.go
@@ -1,6 +1,6 @@
 package server
 
-import "github.com/Peersyst/xrpl-go/xrpl/model/client/common"
+import "github.com/Peersyst/xrpl-go/xrpl/model/requests/common"
 
 type LedgerAcceptResponse struct {
 	LedgerCurrentIndex common.LedgerIndex `json:"ledger_current_index"`
