+++ xrpl/model/requests/clio/ledger_request.go
@@ -3,7 +3,7 @@ package clio
 import (
 	"encoding/json"
 
-	"github.com/Peersyst/xrpl-go/xrpl/model/client/common"
+	"github.com/Peersyst/xrpl-go/xrpl/model/requests/common"
 )
 
 type LedgerRequest struct {
