+++ xrpl/model/requests/ledger/ledger_current_response.go
@@ -1,6 +1,6 @@
 package ledger
 
-import "github.com/Peersyst/xrpl-go/xrpl/model/client/common"
+import "github.com/Peersyst/xrpl-go/xrpl/model/requests/common"
 
 type LedgerCurrentResponse struct {
 	LedgerCurrentIndex common.LedgerIndex `json:"ledger_current_index"`
