+++ xrpl/model/requests/ledger/ledger_entry_request.go
@@ -3,7 +3,7 @@ package ledger
 import (
 	"encoding/json"
 
-	"github.com/Peersyst/xrpl-go/xrpl/model/client/common"
+	"github.com/Peersyst/xrpl-go/xrpl/model/requests/common"
 	"github.com/Peersyst/xrpl-go/xrpl/model/transactions/types"
 )
