+++ xrpl/model/requests/ledger/ledger_entry_response.go
@@ -3,8 +3,8 @@ package ledger
 import (
 	"encoding/json"
 
-	"github.com/Peersyst/xrpl-go/xrpl/model/client/common"
 	"github.com/Peersyst/xrpl-go/xrpl/model/ledger"
+	"github.com/Peersyst/xrpl-go/xrpl/model/requests/common"
 )
 
 type LedgerEntryResponse struct {
