+++ xrpl/model/requests/ledger/ledger_entry_test.go
@@ -3,8 +3,8 @@ package ledger
 import (
 	"testing"
 
-	"github.com/Peersyst/xrpl-go/xrpl/model/client/common"
 	"github.com/Peersyst/xrpl-go/xrpl/model/ledger"
+	"github.com/Peersyst/xrpl-go/xrpl/model/requests/common"
 	"github.com/Peersyst/xrpl-go/xrpl/model/transactions/types"
 	"github.com/Peersyst/xrpl-go/xrpl/test"
 )
