+++ xrpl/model/requests/ledger/ledger_response.go
@@ -3,8 +3,8 @@ package ledger
 import (
 	"encoding/json"
 
-	"github.com/Peersyst/xrpl-go/xrpl/model/client/common"
 	"github.com/Peersyst/xrpl-go/xrpl/model/ledger"
+	"github.com/Peersyst/xrpl-go/xrpl/model/requests/common"
 	"github.com/Peersyst/xrpl-go/xrpl/model/transactions"
 	"github.com/Peersyst/xrpl-go/xrpl/model/transactions/types"
 )
