+++ xrpl/model/requests/path/book_offers_response.go
@@ -3,8 +3,8 @@ package path
 import (
 	"encoding/json"
 
-	"github.com/Peersyst/xrpl-go/xrpl/model/client/common"
 	"github.com/Peersyst/xrpl-go/xrpl/model/ledger"
+	"github.com/Peersyst/xrpl-go/xrpl/model/requests/common"
 	"github.com/Peersyst/xrpl-go/xrpl/model/transactions/types"
 )
