+++ xrpl/model/requests/path/deposit_authorized_response.go
@@ -1,7 +1,7 @@
 package path
 
 import (
-	"github.com/Peersyst/xrpl-go/xrpl/model/client/common"
+	"github.com/Peersyst/xrpl-go/xrpl/model/requests/common"
 	"github.com/Peersyst/xrpl-go/xrpl/model/transactions/types"
 )
