+++ xrpl/model/requests/server/fee_response.go
@@ -1,7 +1,7 @@
 package server
 
 import (
-	"github.com/Peersyst/xrpl-go/xrpl/model/client/common"
+	"github.com/Peersyst/xrpl-go/xrpl/model/requests/common"
 	"github.com/Peersyst/xrpl-go/xrpl/model/transactions/types"
 )
