+++ xrpl/model/requests/subscription/stream/peer.go
@@ -1,6 +1,6 @@
 package stream
 
-import "github.com/Peersyst/xrpl-go/xrpl/model/client/common"
+import "github.com/Peersyst/xrpl-go/xrpl/model/requests/common"
 
 type PeerStatusStream struct {
 	Type           StreamType         `json:"type"`
