+++ xrpl/model/requests/subscription/stream/transaction.go
@@ -3,7 +3,7 @@ package stream
 import (
 	"encoding/json"
 
-	"github.com/Peersyst/xrpl-go/xrpl/model/client/common"
+	"github.com/Peersyst/xrpl-go/xrpl/model/requests/common"
 	"github.com/Peersyst/xrpl-go/xrpl/model/transactions"
 )
